requirements.txt
mysql_connector_repackaged==0.3.1
plotly==5.24.1
yfinance==0.2.50
